Refactor more creation/upgrading into module
